이름변경/회원가입 기능 구현
